Speed up the package
Use parallel computations for slow part of the code to speed up the runtim time for very large BAM files (100 million reads). E.g. we use mclapply to import the BAM file faster: genomic tiles are processed in parallel. Other parts of the code are parallelized with mcmapply. Rcpp is used to speed up the filtering and processing of large data.tables with SJ read pairs.
This commit contains all changes to the code to make it run faster and to use Rcpp in the package.